Start with screens and backstacks from intents
This enables the use of the Intent an Activity is
started with to let Triad start with given screens.
In order for this to work, the Screen class must
have one of the following (in order of priority):

- Either a parameterless constructor
- A static function create(Intent)
- A top level function create(Intent) in a file called "${screen}Kt".

The Intent passed to the create(Intent) function is the
same Intent used to start the Activity.

For example, in a file called MyScreen.kt:

fun create(intent: Intent) = MyScreen("Name")

class MyScreen(val name: String) : Screen<ApplicationComponent>() {
  /* ... */
}