Clear backstack after closing activity with back button
This seems like reasonable behavior. After all, the
activity is cleared and any UI state is restarted.
The next time the activity is started (with Triad still
in memory), `startWith()` will be called again.

This especially fixes an issue where a custom startup
animation would not run in certain cases.